+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Here we write upgrading notes for brands. It's a team effort to make them as straightforward as possible.
 
+### Fixed
+- Using IResponse & IMessage interfaces in response
+- Add more information for response like  is-successful, message, info  & status code
 
 ## [0.1.5-beta] - 2022-06-20
 ### Removed